fix: Point youtube urls to newly uploaded urls
4 YouTube videos in the edx demo course are no longer available to watch
because the YouTube links don't work (Google made the videos private
back in July and the account hosting the videos has been lost). 3 out
of the 4 videos have been downloaded from AWS and re-uploaded to the
openedx YouTube account. The course has been updated with the correct
urls. In addition, srt.sjson files names have been changed to reflect
the change in the YouTube video urls.